// frontend/src/app/core/api/model/payrollRecordResponse.ts
/**
 * Le Tacos Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PayrollRecordResponseDto { 
    id?: string;
    employeeId?: string;
    employeeName?: string;
    amount?: number;
    currency?: string;
    paymentDate?: string;
    periodStart?: string;
    periodEnd?: string;
    paymentMethod?: PayrollRecordResponseDto.PaymentMethodEnum;
    notes?: string;
    transactionId?: string;
}
export namespace PayrollRecordResponseDto {
    export const PaymentMethodEnum = {
        Cash: 'CASH',
        Card: 'CARD',
        Wolt: 'WOLT',
        BankTransfer: 'BANK_TRANSFER',
        Other: 'OTHER'
    } as const;
    export type PaymentMethodEnum = typeof PaymentMethodEnum[keyof typeof PaymentMethodEnum];
}